More clippy fixes

+++ examples/window_debug.rs
@@ -17,7 +17,7 @@ fn main() {
 
     let window = WindowBuilder::new()
         .with_title("A fantastic window!")
-        .with_inner_size(LogicalSize::new(100 as f32, 100 as f32))
+        .with_inner_size(LogicalSize::new(100.0, 100.0))
         .build(&event_loop)
         .unwrap();
